Move socket.h definitions into their own .c file
